instancetype: Reject Matcher Name updates without updating RevisionName
As set out in issue #9132 requests to update the Name of an
{Instancetype,Preference}Matcher were always accepted regardless of the
RevisionName being cleared or updated. This would result in an older and
likely unwanted ControllerRevision being used containing an outdated or
completely wrong Instancetype or Preference.

With this change the VM mutation webhook will now reject requests to
update the target Name of a {Instancetype,Preference}Matcher without
also clearing or updating the RevisionName.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>