e2e: treat kubevirt CR conditions when preparing the environment
The ensureKubevirtInfra() ensures that all kubevirt components
are synced and ready for testing. Virt-operator does the same, and
reports the status of Kubevirt via the Kubevirt CR conditions. So it
doesn't makes sense to compete virt-operator or to mimic its logic.
Its enough to observe the Kubevirt CR conditions, and this change
suggests exactly that.

Signed-off-by: enp0s3 <ibezukh@redhat.com>