e2e test: Fix imagePullSecrets check
Currently, imagePullSecrets tests checks that the virt-launcher pod has no
imagePullSecrets. This is not always true, because if the ServiceAccount that
runs the pod has an imagePullSecret, the pod will inherit it, causing a failure.
This patch changes the check:
If the Kubevirt CR has no imagePullSecrets, then the virt-launcher pod should have
the same imagePullSecrets of its ServiceAccount.

Signed-off-by: fossedihelm <ffossemo@redhat.com>