rpm: Manually fix URL for passt package
The URL generated by 'make rpm-deps' contains a caret, which
bazel loudly complains about at build time:

  Error downloading passt-0^20221110.g4129764-1.el9.x86_64.rpm: Illegal character in path at index 78:
  https://mirror.stream.centos.org/9-stream/AppStream/x86_64/os/Packages/passt-0^20221110.g4129764-1.el9.x86_64.rpm

For whatever reason, this was not a problem with the previous URL
for the passt package, which also contained a caret. In that case,
the unescaped caret only appeared after one that was properly
escaped, which apparently made things okay somehow?

In any case, escaping the caret makes bazel happy and this manual
tweak successfully survives a 'make rpm-deps' run, so we can work
around the problem this way until a proper solution is found.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>