Test: Fix storage lane
Test "[Serial]without fsgroup support should succesfully start"
does need to use a privileged Pod with a namespace. This mean
the namespace sets if test suite can create a privileged pod
when we are running with PSA enabled.

This commit is just moving the pvc and vmi into privileged
namespace.

Signed-off-by: L. Pivarc <lpivarc@redhat.com>