Revert "Add checks of dedicatedCPUPlacement to validation webhook"
This reverts commit 48a2adea3eec3aa267fdaa7139c8e74ca06fb07a.

Thanks to e7898af39b54298043adef02ae578d825743120e [1][2] we now apply
defaults to a copy of the VirtualMachineInstanceSpec before running the
associated validations in the VirtualMachine validation webhook. This
actually resolves the original issue with enabling dedicatedCPUPlacement
from instance types [3] as it ensures any guest visible memory is
expressed as resources requests before we attempt to validate.

As such we can drop the blocking validations from the instance type
webhooks and allow `dedicatedCPUPlacement` to be set once again.

[1] https://github.com/kubevirt/kubevirt/pull/9103
[2] https://github.com/kubevirt/kubevirt/issues/9102
[3] https://github.com/kubevirt/kubevirt/issues/8867

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>