e2e: improve vmpool tests
In case a spec is testing a running VM pool we would like to
make sure VMs are ready. For that purpose we would use ReadyReplicas.

In addition, when scale in the pool we would like to track only the VMs
that are owned by the pool that is under test, this way we can be more
precise in our assertions.

Signed-off-by: enp0s3 <ibezukh@redhat.com>