Remove hostname limitation for migrations
For a long time Libvirt was using hostname
to perform check if it performs migration
from&to the same node. This check was dropped
or rather it is implemented in another way
since 6799b72d927015db4ce4cab879f072abc91a41ae 2020.

Therefore we can drop this limitation and actually
address long standing bug with services.

Signed-off-by: L. Pivarc <lpivarc@redhat.com>