host-disk: remove SELinux categories during migration
Migrating a VMI with a RWX disk requires both the source and target virt-launchers to have access to the disk.
That is not possible when SELinux categories are used inside the PVC.
This commit removes the categories at the beginning of the migration,
then sets them to the correct values at the end of the migration.

Signed-off-by: Jed Lejosne <jed@redhat.com>