[e2e] Fix flaky tls test
This test wants to make sure that the migration proxy server is
protected by tls. It requires that the migration is slow enough
to create a connection with self-signed certificate.
If the migration lasts for a very short time, the connection returns `EOF`
instead of `tls: bad certificate` because the migration ports are already gone.
This fix limits the migration bandwidth, slowing down the migration more than
only run stress-ng.

Signed-off-by: fossedihelm <ffossemo@redhat.com>