Only update the size of existing iso files
Currently, all volumes have an iso file associated with them. But that
will change in the next commits, so some volumes may not have an
associated disk device (e.g., those accessed using a virtiofs
Filesystem), so they do not have an iso file that its size requires to
be updated.

Signed-off-by: German Maglione <gmaglione@redhat.com>