[e2e] Add labels parameter to CreateFSPVC helper function
`CreateFSPVC` is a helper function that creates a PVC.
If you want to add some labels to the PVC, you are forced to
Update the PVC just after the creation. This can cause some
collisions during update because the object has been modified.
Adding the `labels` parameter allows us to save an `Update` and
avoid such collisions described above.

Signed-off-by: fossedihelm <ffossemo@redhat.com>