Create sidecar-shim binary and Container
The idea is to have a base container that can be used to quickly write
sidecars without gRPC boilerplate.

The sidecar-shim binary implements the gRPC boilerplate. The user of
sidecar-shim-image should put his code in $PATH with the name of the
hook he wants to execute (e.g: /usr/bin/onDefineDomain)

The follow-up commit moves the smbios sidecar to use this image.

https://github.com/kubevirt/kubevirt/pull/6975
Signed-off-by: Victor Toso <victortoso@redhat.com>