Create iso files only if Disk is present
Volumes such as confgmaps, secrets, etc. can be accessed using a
virtiofs filesystem. So, in this case, creating an iso file is
unnecessary because it may not have an associated disk device,
for instance:
```
...
devices:
  filesystems:
    - name: my-config
      virtiofs: {}
...
volumes:
  - configMap:
      name: config-vm
  name: my-config
...
```
Let's add a function that only creates iso files only if a disk device
is present it will be used for config maps, secrets, downward api and
service account resources also reducing the duplication of code.

Signed-off-by: Javier Cano Cano <jcanocan@redhat.com>
Signed-off-by: German Maglione <gmaglione@redhat.com>