Dont flood logs on sync guest time
Currently it may happen that we invoke
the sync operation too soon. (Read the domain
is actually not running on the target.) In this
case but also in some case of failure we flood
the logs with warnings.

This commits changes the verbosity of repeating
logs and make sure it is logged at the end.

Signed-off-by: L. Pivarc <lpivarc@redhat.com>