Calculate hotplug attachment pod requests/limits based on VMI ratio
If the VMI ratio of either cpu or memory is 1, then the matching
type of ratio for the hotplug attachment pod will also be 1.
Higher ratios are possible and will be calculated accordingly. To
ensure ratios are always <= a limitRange, the value is rounded up
to the nearest int64 which ensure the limitRange validation succeeds

Signed-off-by: Alexander Wels <awels@redhat.com>