Add chroot capability to rsync container in hack/dockerized
The latest version of podman (>v4.4.0) has dropped the CHROOT capability
from the default configuration[1]. The rsync container used as part of
cluster-sync requires chroot capabilities so this causes it to fail.[2]

The update allows it to run successfully with podman >v4.4.0.

Also tested the docker flow and this change has no impact on users
running with docker.

[1] https://blog.podman.io/2022/12/dropping-capabilities-making-containers-more-secure/
[2] https://github.com/kubevirt/kubevirt/issues/9279

Signed-off-by: Brian Carey <bcarey@redhat.com>