sidecar: disk-mutation: change base image to distroless/base
Current base image is Fedora 31 which is EOL since 2020-11-24.
Let's move the base image to a base image that we use and update more
often across components.

Follow the same approach as cmd/virt-api to have an intermediate
version-container image for the base.

See: https://github.com/kubevirt/kubevirt/pull/9229
Signed-off-by: Victor Toso <victortoso@redhat.com>