Add AdditionalGuestMemoryOverheadRatio to Kubevirt CR
AdditionalGuestMemoryOverheadRatio can be used to set a
ratio between VM's memory guest and the memory allocated
to the compute container. A higher ratio means that the
VMs would be less compromised by node pressures, but
would mean that fewer VMs could be scheduled to a node.
If not set, the default is 1.

Signed-off-by: Itamar Holder <iholder@redhat.com>