[e2e] Fine-tuning of the pausing test
Add 4 seconds of toleration from the comparison between the guest and the host uptime difference
before pause.
The uptime command returns the value in seconds.
If you capture the uptime when it is at the beginning of that second or at the end of that second,
the value comes out the same even though in fact a whole second has almost passed.
In extreme cases, as we take 4 readings (2 initially and 2 after the unpause), the deviation could be up to just under 4 seconds.
This fact does not invalidate the purpose of the test, which is to prove that during the pause the vmi is actually paused.

Signed-off-by: fossedihelm <ffossemo@redhat.com>