Use virtio-scsi as the default storage controller
Virtio-blk disks consume addresses on the PCI bus and this is an important limitation. Since performance parity has been shown between virtio-scsi and virtio-blk it makes sense to standardize the former.

This commit updates the mutating webhook for VMs/VMIs so, by default, disks are created using SCSI bus when possible.

Signed-off-by: Alvaro Romero <alromero@redhat.com>