node-labeller: Check arch on the handler side
In a cluster with mixed x86 and arm workloads, the operator job
deployment can not generate the daemonset spec properly for both x86 and
arm handler nodes, as it has only one `runtime.GOARCH`.

So it's better to create the init container for all handler pods and do
the arch check in node-labeller.sh instead.

Signed-off-by: Li Yuxuan <liyuxuan.darfux@bytedance.com>