test.sh: Skip sleep for kind and k3d providers
the sleep was introduced in order to reduce stress on disk
in case a lot of jobs start on the same time.

Since there is only one SR-IOV at a time, it is safe to let
it start without delay, because it won't create stress.
The rest of the non SR-IOV jobs will have this sleep.

It will save us time while running SR-IOV jobs, which have very
limited resources.
Same for kind vGPU, since it runs only one at a time.

Signed-off-by: Or Shoval <oshoval@redhat.com>