e2e, network, Expose: Reuse VMIs
In the Expose tests suite there are tests batches that create the exact same
VMI before each test but assert on diffrent things regardless of the VMI spec.

Creating the same VMI over and over is which is not efficient and time consuming
(VMs takes 30~60 seconds to boot, multiply it with the tests count..).

With the combination of the following decorators it is now possible to reuse VMIs
in multiple tests and shorten the overall time of the test suite:
- Ordered
  Set a tests-container (e.g: Context/Describe) to run test specs in the order
  they are written and enable using BeforeAll [1] [2]
- ContinueOnFailure
  Sets a tests-container to tolerate failing tests and continue [3]
- RetainVirtualMachineInstances
  Make the global test suite clean-up to skip VMIs and virt-launcher pods.

[1] https://onsi.github.io/ginkgo/#ordered-containers
[2] https://onsi.github.io/ginkgo/#setup-in-ordered-containers-beforeall-and-afterall
[3] https://onsi.github.io/ginkgo/#failure-handling-in-ordered-containers

Signed-off-by: Or Mergi <ormergi@redhat.com>