e2e: fix flakiness vm foreground finalizer test
The AfterEach function deletes the vm created in the It spec.
This is used from two tests; in one of them the deletion is already performed,
but the vm does not disappear since it has a custom finalizer.
In AfterEach, before the vm deletion, the custom finalizer is removed from the vm.
Since the deletion was already fired the subsequent deletion may fail due to:
`Message: "virtualmachines.kubevirt.io \"testvmi-hcx9l\" not found",`
This fix prevents this flakiness by ignoring the "Not found" error.

Signed-off-by: fossedihelm <ffossemo@redhat.com>