fix out-of-bounds slice access in evacuation controller
There can be a case where there are no free spots but also
there are no migration candidates and no non-migratable VMs.
in such case the execution continues. This could lead to OOB
access, since the free spots counter can have a negative value.
A negative value can be reached if for example a user decides to lower
the configured maximum of allowed concurrent migrations per source node.
Thus the amount of active migrations can exceed the maximum and we get a
negative freeSpots

Signed-off-by: enp0s3 <ibezukh@redhat.com>