tests: Drop createdVirtualMachine variable
Previously createdVirtualMachine was used in AfterEach() to cleanup the
VM. However, the var was not properly initialized in one of the tests
and that was causing failures when deleting the object. Such a cleanup
is not needed in general, since VMs are deleted during global cleanup
after each test.

Also ensure the other variables used for the cleanup are properly
null'ed in AfterEach().

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>