Workaround for nonRoot FG on Kubevirt (#2273)
With https://github.com/kubevirt/kubevirt/pull/8563
Kubevirt introduced a new FG named `Root`,
the `nonRoot` FG we are using is still there but
declared as deprecated.

Unfortunately due to bug on Kubevirt the now
deprecated `nonRoot` FG we rely on is
completely ignored so, as a workaround, we will
start internally translating nonRoot -> Root
on Kubevirt with negative logic.

In a future PR the nonRoot FG will be properly deprecated
also here and a new Root FG will
be introduced with a proper conversion logic on upgrades.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=2175171
JIRA-ticket: https://issues.redhat.com/browse/CNV-26406

Signed-off-by: stirabos <stirabos@redhat.com>