Check for PVC and DataVolume existence in a central place
Only proceed to template rendering when all PVCs are present.  The
template rendere also checks for PVC existence, but if the PVC appeared
between the check and the template rendering, the limited render-check
will miss subtlies like checking if the PVC belongs to a DV and an
import needs to be finished before creating the launcher pod.

This fixes a race, where sometimes a launcher pod for a VMI could
already be created while the DV was in the middle of creating the PVC
and importing.

Signed-off-by: Roman Mohr <rmohr@google.com>