virt-launcher: add cpu hotplug functionality
add SyncVirtualMachineCPUs to virt-launcher API
exposed functions of libvirt interface that related to VCPU updates.
set the disabled and hotpluggable vcpus based on the MaxSockets
set pinned vcpus in reference to MaxSockets
refactor libvirt live update flags

Define only one socket as not hotpluggable
in order to be always able to scale down to
topology with 1 socket.

Signed-off-by: Vladik Romanovsky <vromanso@redhat.com>
Co-authored-by: enp0s3 <ibezukh@redhat.com>
Co-authored-by: Luboslav Pivarc <lpivarc@redhat.com>