Disable PSA for upgrades with kv v0.58 on OCP/OKD (#2281)
Disable PSA enforce=restricted to workaround
upgrading from kv v0.58 on OCP/OKD 4.13.
OCP/OKD are supposed to be shipped without setting
PSA enforce=restricted by default but actual builds
are still enforcing it.

Kubevirt v0.59.0 is PSA compatible but
Kubevirt v0.58.z is not so the upgrade-prev
ci lanes are going to fail on OCP/OKD 4.13
due to this.

Temporary explicitly opting out from
enforce=restricted as a workaround.

Signed-off-by: stirabos <stirabos@redhat.com>