A logging convention
We currently use numbers from 0-9 range to define how verbose we want
our log to be. This commits introduces the 8 predefined CONSTs to be
used across multiple components and make it clear the verbosity level
expected for each message.

As `-v 2` is the default in k8s, 0 and 1 are named to accordantly as
'ALWAYS' and 'HIGH', respectively follow by 'DEFAULT' for 2. The
higher log levels are meant to be used for troubleshooting.

'DEBUG' is expected to be more used in scenarios that can aid
developers. 'FIXME' in meant to non error type of issues that must be
addressed eventually.

'TRACE' and higher verbosity levels are more useful in non stable
components (e.g: Drivers for multiple and/or untested hardware).

'DUMP' could be used for detailed information of memory/protocol, etc.

The 7, 8 values were left blank for discussion.

'RARE' is meant to not be shown easily. It can also be used to fine
tune sub-components (e.g: could show libvirt logs even if
libvirtLogVerbosityEnvVar was not set).

Signed-off-by: Victor Toso <victortoso@redhat.com>