launcher: nic hotplug based on the presence of the iface in the domain
We were currently hotplugging interfaces based on the interfaces present
in the `VMI.Spec.Interfaces`, but not present in the
`VMI.Status.Interfaces`, having the multus source reported in the
interface info source.

This commit changes this, by hotplugging based on interfaces present in
the `VMI.Spec.Interfaces`, but not presently available in the
**domain**, which is more accurate.

As a side-effect, we no longer calculate the new domain interfaces in
the `nichotplug.go` file, since the `updatedDomain` already features
them - it was redundant. Removing this calculation makes the code more
readable, since the domain is not mutated in the hotplug function.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>