api: Add RuntimeClassName field to VirtualMachineInstanceSpec
To set a VMI with high performance utilities such as DPDK and Realtime,
the containers may nee to run with elevated privileges. The way this is done
is by setting the appropriate runtimeClass to the pod [0].

Currently a VMI cannot use a specific runtimeClassName on the
virt-launcher pod, unless that is configured for all the VMIs on the
cluster [1].
In order to allow configuring specific VMIs to different
runtimeCladdName, adding the RuntimeClassName field to
VirtualMachineInstanceSpec.

[0]
https://docs.openshift.com/container-platform/4.12/scalability_and_performance/cnf-low-latency-tuning.html#node-tuning-operator-disabling-cpu-load-balancing-for-dpdk_cnf-master

[1]
https://github.com/kubevirt/kubevirt/blob/2fa56e8baf4e16bd2a0b8660b05add1108d3365a/pkg/virt-controller/services/template.go#L544

Signed-off-by: Ram Lavi <ralavi@redhat.com>