virt-api, iface-hotplug: Drop redundant conditions
The `request.AddInterfaceOptions != nil` has been used in a defensive
manner on functions which are dedicated to adding an vNIC.
I.e. the code flow would have reached these functions only if the
operation has been recognized as a hotplug, implying that the request
has this data and there is no need to re-check it again in the stack.

Signed-off-by: Edward Haas <edwardh@redhat.com>