network: Refactor network handling to integrate hotplug
Embed most of the hotplug logic into the existing network setup flow.

This includes the logic of identifying if hotplug should be done and on
which networks. Functions signatures have been adjusted to fit both the
regular and hotplug setups.

The change handles both phase 1 and phase 2.

Future changes should strive to unify the logic, such that even a
regular network setup can reconcile by hotplugging the missing
networks.

Signed-off-by: Edward Haas <edwardh@redhat.com>