instancetype: Handle DataVolumeSourceRef without namespace when inferring defaults
As set out in issue #9431 it is valid for a DataVolumeSourceRef to not
provide a namespace, the assumption being that the resource it points to
is in the same namespace as the VirtualMachine.

This change handles this by switching to the namespace of the
VirtualMachine when no namespace is provided by the DataVolumeSourceRef,
avoiding a previous panic.

NOTE(lyarwood): conflict caused by
c09acad4db9cddf75c5b059d62c1074c40bd85ab not being present the in
release-0.59 branch.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>
(cherry picked from commit 50befd593b10b6a5c2a574f12ff9c6cbd6a9674d)