instancetype: Handle DataVolumeSourceRef without namespace when inferring defaults
As set out in issue #9431 it is valid for a DataVolumeSourceRef to not
provide a namespace, the assumption being that the resource it points to
is in the same namespace as the VirtualMachine.

This change handles this by switching to the namespace of the
VirtualMachine when no namespace is provided by the DataVolumeSourceRef,
avoiding a previous panic.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>