webhooks: Return early when VirtualMachine is being deleted
As set out in issue #9438 previously attempts to remove a namespace
would be blocked if it contained VirtualMachines using
VirtualMachineInstancetypes. This was due to the update call to remove
the virtualMachineControllerFinalizer being rejected as the
ControllerRevision associated with the VirtualMachineInstancetype had
already been deleted, causing admission of the request to be rejected.

This change adds basic logic to both the mutation and admission webhooks
that returns early *if* the VirtualMachine is being deleted. Ensuring
the update to remove the virtualMachineControllerFinalizer is successful
and allowing the overall namespace removal to also complete.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>