e2e: improve the kubevirt health-check function
When updating the Kubevirt CR, It takes time for the virt-operator
to process the new desired state and report the status.
If the HC function called in that time-frame we may think Kubevirt is synced.
Therefore I've added another check, that compares the generation numbers.
Its equity is an evidence that the execution loop of virt-operator is finished.

Another improvement is to check whether virt-operator is alive. Since
KV CR is owned by a user rather than by virt-operator, if operator was
deleted somehow, the CR is left and not cleaned by GC, and the status
information there is not relevant.

Also, health-check added to the initial before suite stage (SynchronizedBeforeTestSetup)

Signed-off-by: enp0s3 <ibezukh@redhat.com>