network, api: Replace `podConfigDone` field with `infoSource`
Network detected on the multus networks status pod annotation has been
marked on the VMI status interface using a dedicated flag
(`podConfigDone`).

This change removed the flag in favor of a new source option in the
`inforSource` field.

The virt-controller detects a multus network status on the VMI pod and
adds the `multus-status` source to the `infoSource` field.
The virt-handler, when calculating the VMI status interfaces, preserves
this information.

Signed-off-by: Edward Haas <edwardh@redhat.com>