virt-operator: fix updated canary pod detection in virt-handler rollout
virt-operator uses canary upgrade method when rolling out the virt-handler. A pod
considered canary pod if its the first one to be ready after the roll-out, if that
pod is ready, virt-operator speeds up the rollout process by setting maxUnavailable:10%
at the rolling strategy in virt-handler daemonset spec.

The issue is that in order to determine whether a particular virt-handler pod is updated
the PodIsUptodate() is used, which is not enough because it only tells if a pod is created by
the target install strategy. It is helpful only in kubevirt upgrade process, but not in case
of update of specific virt component. Therefore another indication needed. Using generation
numbers wasn't reliable because early comparison of these numbers could lead to false conculsion
that handler rollout has finished. Daemonset status also wasn't enough because the ready counter
aggregates both pods with new and old specs.

The most precise method was to calulate the hash of the desired spec, and to compare it with
the value of controller-revision-hash annotation on the daemonset pods.

Therefore I took the exact same hash computation functions from k8s code as they appear in
the k8s damonset controller code.

Signed-off-by: enp0s3 <ibezukh@redhat.com>