hotplug: add migration based tests
Those tests will run only on network lanes that have multus v3.

Since hotplug to a pod was introduced on multus v4, clusters with multus v3
cannot support VM in place hot plug.
When hotplugging an interface to a VM running on a cluster with multus v3,
the new interface will be added to the VM/VMI immediately, but the
virt-launcher pod will be updated only after a migration (when a new
virt-launcher pod will be created).
Only then the hotplug will be completed and the interface will be
hot-plugged to the guest.

Signed-off-by: Alona Paz <alkaplan@redhat.com>