virt-launcher, converter: Refactor CreateDomainInterfaces()
Currently, CreateDomainInterfaces() is called in two scenarios:
1. Regular VM setup.
2. NIC hot-plug.

The function currently receives a variadic parameter [1] that
represents the VM's logical interfaces that should to be plugged.

The function also have an internal logic to determine whether
it was called from a setup or hot-plug scenario.

Adjust the function signature to fit both scenarios.
Remove the above internal logic, similarly to what was done in PR #9406.

[1] https://go.dev/ref/spec#Function_types

Signed-off-by: Orel Misan <omisan@redhat.com>