Decrease portForward timeout during WaitForConfigToBePropagatedToComponent
Currently, when we update the kubevirtCR (in serial tests or in AfterEach function),
we ensure that the config is propagated checking the revision number of the
components. To do this, we do a port forward to the pods.
This port-forward has a timeout of 10 seconds, and it is inside an Eventually
block with max duration of 10 seconds and 1 second of polling interval.
Since the next polling in the Eventually block does not start until the first one
is finished, we only try once to do the port-forward.
Let's try at least 2 times, decreasing the port-forward timeout to 5 seconds.

Signed-off-by: fossedihelm <ffossemo@redhat.com>