tests: remove test for IO error with faulty disks
In the storage test suite, there are 2 tests for verifing IO errors. One
of the two tests only checks that the VM is paused on an IO error, while
the second covers also the recovery from the error.

The removed test uses dmsetup to create a faulty disk and it was
introduced before relying on the scsi_debug module.

The 2 test creates flakiness as they both run in parallel and create a
local PV/PVC couple with the same storage class with the function
CreatePVandPVCwithFaultyDisk. For semplicity we can remove the simplest
test.

Signed-off-by: Alice Frosi <afrosi@redhat.com>