Let virt-handler fill the qemu machine type into the VMI status
The libvirt domain XML contains the final qemu machine type. Propagate
that information to the VMI status.

This helps identifying the final machine type when an alias like `q35`
is configured on the VMI spec, which in turn helps finding out if the
aliased machine type from an older release is live-migration-compatible
with the new machine type on a newer release.

Signed-off-by: Roman Mohr <rmohr@google.com>