e2e, SR-IOV: Ignore SyncVMI timeout warning
Since we run on the SR-IOV machine more non SR-IOV
jobs in parallel, the machine might be busy
and have timeouts of SyncVMI.
Those are transient warnings which have a built-in retry
mechanism, as was seen before.
Add another variant of the warning string that should be ignored.

For more info see https://github.com/kubevirt/kubevirt/issues/5027

Signed-off-by: Or Shoval <oshoval@redhat.com>