api, AddInterfaceOptions: Rename NetworkName
PR #6852 introduced the AddInterfaceOptions struct,
which is used as part of a NIC hot-plug.

Currently, this struct contains a field named `NetworkName`.
The field name is misleading, because it represents the name of a
NetworkAttachmentDefinition[1] object, and not a Network[2] object.

Rename this field so its name will properly represent its value.

[1] https://github.com/k8snetworkplumbingwg/multus-cni/blob/master/docs/how-to-use.md#create-network-attachment-definition
[2] http://kubevirt.io/api-reference/main/definitions.html#_v1_network

Signed-off-by: Orel Misan <omisan@redhat.com>