Fix seccomp profile test
The installation test of the custom seccomp profile acts like this:
1. delete previously profile in the virt-handler pod, if exists.
2. enable the feature gate in order to install the profile.
3. check if the profile exists in the virt-handler pod.

To enable the feature gate it is used the utility function
`tests.EnableFeatureGate(virtconfig.KubevirtSeccompProfile)`
The latter does not perform any update if the feature is already present.
Since the feature is enabled by default, this results in a non action, and
the profile deleted at point 1. will not be re-installed, causing the failure.
To effectively check that, once you add the feature gate, the profile is
installed, firstly delete the feature gate (if present), and then add it again

Signed-off-by: fossedihelm <ffossemo@redhat.com>