tests, console: Prevent NewExpecter negative timeout scenario
The NewExpecter constructor function is creating a serialConsole object
and then spawns a generic process - both of which are included for the
given NewExpecter timeout.
Sometimes the first action takes longer than the whole timeout, causing
the spawned process to start with no chance to finish within the
overall timeout duration.
In order to fix this, Adding a check to prevent the spawned process from
starting if the timeout has already passed.

Signed-off-by: Ram Lavi <ralavi@redhat.com>