[e2e tests]: foreground delete testing manifests
Sometimes we pick up the wrong disks-images-provider pod due to
the deletion happening asynchronously. It may be nicer to just
make sure the deletion of the pods is done here.
This appeared to be the motivation in https://github.com/kubevirt/kubevirt/pull/2182
(Not sure why background was chosen, as that is the default)

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>