virt-launcher, manager: Simplify interfaces to hotplug realization
Currently, networksToHotplugWhoseInterfacesAreNotInTheDomain() returns a
VMI spec.network slice to fit SetupPodNetworkPhase2(), and then use that
network object to get the domain interface object.

Eventually, only the VMI spec.network / status.interface name (same thing)
is needed to get the right VMI spec.network and the domain interface.

Change the logic of hotplugVirtioInterface() to first realize the names of the
interfaces to hotplug, and then use it to fetch the VMI status.interface and
domain interface objects.

Signed-off-by: Or Mergi <ormergi@redhat.com>