DownwardMetrics: Refactor scraper
In the following commits the virtio-serial "transport" for
downwardmetrics will be added. Currently, metrics are provided to the
VM using a disk (which will be deprecated in a future release).

The `Report()` method not only collects the metrics, but also writes
them to disk. To facilitate code reuse we will refactor `Report()` to
return the metrics instead of writing them to disk. Keeping `Scrape()`
as is, since it is part of an interface also used by prometheus.

Signed-off-by: German Maglione <gmaglione@redhat.com>
Signed-off-by: Javier Cano Cano <jcanocan@redhat.com>