Enable freePageReporting for non-high performance vmis
With freePageReporting[1] the guest OS informs the hypervisor[2]
about pages which are not in use by the guest anymore.
The hypervisor can use this information for freeing these pages.
`freePageReporting` is an attribute that can be defined at
Memory balloon device in libvirt[3].

Currently, freePageReporting for the memballoon device is disabled
by default, since in libvirt you need to explicitly enable it.

With this, freePageReporting will be enabled by default for all the
new vmis which does not have `autoattachMemballoon` equal to false, AND
which are not requesting any high performance components.
A vmi is considered as high performance if one of the following is true:
- the vmi requests a dedicated cpu.
- the realtime flag is enabled.
- the vmi requests hugepages.

vmi creators can still opt-out from freePageReporting adding the annotation
`kubevirt.io/free-page-reporting-disabled: "true"` to the vmi spec.

[1]https://docs.kernel.org/mm/free_page_reporting.html
[2]https://lwn.net/Articles/808807/
[3]https://libvirt.org/formatdomain.html#memory-balloon-device
Signed-off-by: fossedihelm <ffossemo@redhat.com>