instancetype: Handle VirtualMachineInstancetypeSpecRevision without APIVersion
As set out in issue #9261 there have been reports of older
VirtualMachineInstancetypeSpecRevision objects being stored in
ControllerRevisions without APIVersion set. This is likely due to a
known client-go issue [1] stripping TypeMeta from returned objects.

Given the limited support for v1alpha1 and the fact that we moved away
from VirtualMachineInstancetypeSpecRevision to complete objects being
captured in ControllerRevisions by v1alpah2 this change simply works
around this issue by ignoring the value of APIVersion. This should allow
the conversion to complete correctly until we upgrade these
ControllerRevisions in the future.

[1] https://github.com/kubernetes/client-go/issues/541

Conflicts:
  pkg/instancetype/instancetype_test.go

NOTE:
The conflict was due to refactorings that were not applied to this
branch. (e.g. renaming of variables)

Co-authored-by: Felix Matouschek <fmatouschek@redhat.com>
Signed-off-by: Lee Yarwood <lyarwood@redhat.com>