Improve mountinfo filtering for NFS disks
When a VM was running multiple NFS disks (one boot disk)
and one hotplugged disk. The mount filtering was returning
multiple possible mounts on the host, and sorting on the
length of the 'root' path of the mount. For NFS these are
the same (/). This caused the wrong NFS mount to be picked
when hotplugging and an attempt was made to hotplug the boot
disk. This commit improves the situation by passing the 'source'
field to the filter, and it filters out entries that don't have
matching sources. This allows us to find the correct mount on
the host.

Signed-off-by: Alexander Wels <awels@redhat.com>