node-labeller: Remove pipefail to fix non kvm envs
At some scenarios like running kubevirt on github actions there are not
nested virtualization so kvm is not present at nodes and emulation have
to be used, the node-labeller.sh script was failing if there is no kvm
at pipefail. This change remove pipefail and also add "-x" to show the
script and debug it in case of failure.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>