logs: keep the logger variable type on V()
A followup change in the logger will introduce the
FilteredVerbosityLogger type that is called when calling V() method.

As in the case of non failure, we set the log to verbosity 4 we need that
logger variable keeps the same type. Just call V(2) to have it
properly defined as FilteredVerbosityLogger.

Signed-off-by: Victor Toso <victortoso@redhat.com>