api: Move the core API storage version to v1
This change moves the storage version for all core API CRDs to v1. This
does not impact existing objects that will continue to be stored using
their original v1alpha3 version while being served as v1alpha3 or v1, as
was the case previously.

Work will be required in the future to ensure all stored v1alpha3
objects are read and updated to v1 but for the time being this isn't
required as part of this PR.

For more context please review the following k8s documentation:

https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definition-versioning

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>