virt-operator: fix kubevirt version parsing
Kubevirt upgrade process can be tracked by watching the Kubevirt CR status conditions.
However anotherway to track it is to wait for the `KUBEVIRT_VERSION` value to be equal
to `status.observedKubevirtVersion` in the Kubevirt CR. We've found that the
`status.observedKubevirtVersion` dissapears from the status when providing virt-operator
image name that not ends explicitly with `virt-operator` suffix
(`virt-operator-1` for example).

The suggested change fallbacks the kubevirt version
to `KUBEVIRT_VERSION` env var, and in case the var is not set, it simply fallbacks to `latest`.

Signed-off-by: enp0s3 <ibezukh@redhat.com>