SELinux policy: ensure it never gets installed if the feature gate is set
Most of the virt-handler callbacks are setup before refreshing the cache.
Also, callbacks are called once on setup.
It means that on first time, callbacks get called with mostly unpopulated objects,
including a crucially empty set of feature gates.
Since the SELinux policy feature gate is of the "disable" type,
the policy is then considered enabled and therefore gets installed.

Adding the callback after the cache refresh fixes the issue.

Signed-off-by: Jed Lejosne <jed@redhat.com>