virt-handler, net: Intorduce ConfigStateCache
This commit extends ConfigState to contain in-memory cache alongside with
a file cache.
This is needed to avoid reading from the file each time the cache is
accessed.

Currently this cache is used by SetupPodNetworkPhase1 in virt-launcher's
network namespace.
The next commits will use this cache to check whether virt-handler should
try to plug an interface or not before stepping into the virt-launcher's
network namespace.

Signed-off-by: Alona Paz <alkaplan@redhat.com>