virt-controller, hotplug: Use the same name-scheme as other interfaces
On hotplug interface, check the VM current pod interfaces names
and use the same name-scheme for the interface that is about to
hotplug.

When there are no interfaces plugged to the VM, default name-scheme
is used (hashed names).

NonDefaultMultusNetworksIndexedByIfaceName moved as is to another
package to prevent cyclic dependency between 'watch' and 'services'
packages.

Signed-off-by: Or Mergi <ormergi@redhat.com>