api: Introduce the `kubevirt.io/interface` resource name
Users may express the requested number of network interfaces that
the domain should reserve for future additions (hotplug).

The value is expressed through the VMI spec, as a resource-name
(`kubevirt.io/interface`) in the `resources.requests` list.

The value may be any integer between 0 and 32 (including).

Semantics are not included in the API layer logic, preserving simplicity
and avoiding dependency on existing state.

The new resource is reserved for kubevirt usage only.
virt-controller resource rendering is filtering this new resource in
order not to pass it to k8s entities (i.e. the container).

Signed-off-by: Edward Haas <edwardh@redhat.com>