device-manager: add plugin for USB pass through
tldr: This plugin will handle USB HostDevice configuration introduced
in previous commit. The plugin verifies the request and match with
plugged USB devices in the Node. When successful, it exports the
Devices to k8s and allow allocation said devices to VMs running in
this Node.

A few helpful assumptions that make understanding the whole:

 1. We create one USB Device Plugin per resource name
 2. Each USB Selector defines a single USB device
 3. Each resource name can have multiple USB selectors

Because of (2) and (3), each k8s Device exposed in ListAndWatch() can
actually mean multiple USB devices.

Some examples to clarify the above:

 E1: - resourceName: kubevirt.io/storage
       selectors:
         - vendor: "46f4"
           product: "0001"

     Defines a single USB device per k8s Device. If node01 has 3 such
     USB devices, "kubevirt.io/storage" will expose 3 k8s Devices for
     node01

 E2: - resourceName: kubevirt.io/my-use-case
       selectors:
         - vendor: "dead"
           product: "beef"
         - vendor: "feed"
           product: "faad"

     In case node01 has 3 dead:beef but only 1 feed:faad USB devices,
     "kubevirt.io/my-use-case" will expose 1 k8s Device (with 2 USB
     devices for it) for node01.

Signed-off-by: Victor Toso <victortoso@redhat.com>