docs: Add upgrade note regarding v1 storage version migration
The following PR recently merged moving the storage version of all core
objects to `v1`:

https://github.com/kubevirt/kubevirt/pull/9628

To accommodate the eventual removal of the original `v1alpha3` version
this change adds a note to the updates document recommending the use of
the kube-storage-version-migrator tool to automate the upgrade of all
stored core objects to v1 ahead of the removal of v1alpha3.

An older note suggesting new CRD versions are not allowed is also
removed as this is no longer the case.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>