Add static rate limit profile `highBurst` (#2318)
The rate limit profile `highBurst` enables to create and maintain
a high number of VMs in the same cluster. This profile could be
enabled adding the following to the HCO definition:

```
apiVersion: hco.kubevirt.io/v1beta1
kind: HyperConverged
...
spec:
  tuningPolicy: highBurst
...
```
Also, it adds test, functest and documentation about the profile.

Signed-off-by: Javier Cano Cano <jcanocan@redhat.com>