tests, network: Adjust helper to support multiple secondary nets
The change includes a fix for the network hotplug on VM objects.
With VMI/s, the hotplug is reflected immediately on the manifest,
but for VM/s, the hotplug is reflected on the underlying VMI after some
time. Therefore, there is a need to wait for the spec patching to appear
on the VMI spec and only then proceed.

Signed-off-by: Edward Haas <edwardh@redhat.com>